daemon: ensure newlink for bridge member is removed from snapshot on delete
It has been observed that when deleting an interface without first
changing it to remove it as a member of a bridge/switch then the
dellink message can omit the master device, whereas the updating of
the snapshot database assumed that this sort of change on an
RTM_DELLINK cannot happen.

So to cope with the possibility that previously an interface has an
RTM_NEWLINK messages stored in the snapshot slave_link table, and an
RTM_DELLINK message is seen without a master device and without an
intervening update that would have migrated it from the slave_link
table to the link table, then try to delete from both in the absence
of a master interface attribute.

Fixes: VRVDR-49601