snapshot: add separate levels for team and bridge slaves
When a bonding interface is a switch-member then this results it
simultaneously being both a master (team) and a slave (bridge), and
currently with both team and bridge slaves being in the same snapshot
hash this can then result in the RTM_NEWLINK messages being delivered
in an order such that the RTM_NEWLINK message for a bond-group member
is sent before the RTM_NEWLINK for the bonding interface.

In order to ensure that an RTM_NEWLINK for the bonding (team)
interface is sent before that of the bond-group member, when the
bonding interface is also a bridge/switch member then split the
slave_link hash table into separate bridge_slave_link and
team_slave_link hash tables and replay them in that order to the
dataplane.

Fixes: VRVDR-49763