Don't sort cstore items that are bound to an interface
Application configuration (cstore) messages are stored and published
as they are received. Subsequent replay of the messages (snapshot) is
in application order by virtue of the associated sequence number
assigned when first published.

However if the messages are "bound" to an interface, that has yet to
be defined, they are stored but not published, i.e. there is no
associated sequence number (all have -1). Once the interface is
defined, the associated messages are collected and published. However
the collection process generates the list in reverse order (items are
pushed to the front of the list). The subsequent sequence number sort
further jumbles the list of commands.

When creating the interface message list add messages to the end of
the list, i.e. in order. Further when performing the sort adjust the
compare function result to ensure the order remains unaltered when
processing items with no sequence number.

Fixes: VRVDR-50606