port: set ethtool speed/duplex values to unknown on creating a port
The dataplane only sends LINKUP/LINKDOWN messages to the controller on
transitions, meaning that if an interface is in admin-down state from
bootup then no LINKUP/LINKDOWN message will be sent by the dataplane
for it. The kernel tun/tap driver initialises the ethtool settings to
10Mb/full duplex/Twisted Pair which results in confusing output.

Fix this by setting those values to unknown values on port creation,
in the same way as would be done if the port was subject to an
admin-up/admin-down cycle.

Fixes: VRVDR-45369
(cherry picked from commit ef052ef03a2f2a8d62103ec5b24360c69ffe72e6)