cstore: add an api to store config in a protobuf message
The vplaned lib allows callers to send string based messages to
store in the cstore.  Add an api to store a protobuf based message.

This user provided protobuf message will be wrapped in a
DataplaneEnvelope, then wrapped in a VPlanedEnvelope, then base64
encoded so that it can be passed as a string. It will then be stored
as a protobuf encoded message.

VRVDR-51445