port: delete kernel interfaces for local vplanes
Performing the creation/deletion of kernel interfaces in the
controller has the benefit of being able to purge stale interfaces
from the system after a dataplane restart where the configuration may
have changed in the intervening period, or otherwise the interfaces
aren't created again. It also means that it removes an obstacle to
cleaning up the ifp life-cycle in the dataplane, as well as making the
design easier to understand.

Therefore, delete the kernel interface for a port for local vplanes
and not just remote ones. Also set the dormant state of the interface
since the code already exists to do this and it avoids duplicate code
in the dataplane to do the same job.

VRVDR-51844