vplaned: ignore unresolved mroute netlink updates
An unresolved multicast route will be received as RTM_NEWROUTE but with
an RTA_IIF ifindex of zero. Later, if it is resolved, we will receive a
further RTM_NEWROUTE but with a valid non zero ifindex. However, both
messages have been saved to the snapshot. Hence, on replay, even if the
multicast route has been deleted, we will recreate the unresolved route.
Either we remove this unresolved route from the snapshot when we receive
an update, or we just ignore it. As I don't think the dataplane cares
about unresolved routes, as there is nothing to forward, I think we can
just ignore the unresolved message.

Fixes: VRVDR-47352