Update "snapshot" command to be able to display cstore keys
Running the "snapshot -c" command is used to request that the
configuration information held in the cstore be displayed.  However, it
is not currently possible to display the database key information. Often
there are bugs relating to incorrect key information being used
(e.g. containing non-key information or missing a key).  This updates the
code to return the database key information to the client.  which will
be shown if the "-d" (debug) option is provided along with "-c".

VRVDR-56240